refactor(visualization): split codeCharta.model into domain + state behind a barrel
Step 1 of incrementally breaking up the 363-line model god-file. Definitions
now live in model/domain.model.ts (leaf .cc.json domain types) and
model/state.model.ts (the ngrx state tree, which imports the domain one-way).
codeCharta.model.ts becomes a pure barrel re-exporting both, so all 416
existing importers keep working unchanged. Consumers can migrate to the
specific files over time, then the barrel can be dropped.

Verified: dependency-cruiser 0 errors / 94 warnings / no new cycles, tsc 0
errors across all importers, model tests green.
